pyproject.toml
[project]
name = "multi-tenant-github-mcp"
version = "0.1.0"
description = "multi-tenant-github-mcp MCP server built with FastMCP"
authors = [
    {name = "ceposta", email = "christian@solo.io"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp @ file:///Users/christian.posta/python/mcp-python-sdk",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "fastapi>=0.116.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project.scripts]
"multi-tenant-github-mcp-server" = "src.main:main"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=22.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
    "types-PyYAML>=6.0.0",
]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true 
